ci: split some data files to be smaller
