+++ lib/supporting/RewardData.js
@@ -1,0 +1,79 @@
+'use strict';
+
+const resources = [
+  'Alloy Plate',
+  'Argon Crystal',
+  'Circuits',
+  'Control Module',
+  'Cryotic',
+  'Ferrite',
+  'Gallium',
+  'Morphics',
+  'Nano Spores',
+  'Neural Sensors',
+  'Neurodes',
+  'Orokin Cell',
+  'Oxium',
+  'Plastids',
+  'Polymer Bundle',
+  'Rubedo',
+  'Salvage',
+  'Tellurium',
+];
+
+// Aura names
+const auras = [
+  'Brief Respite',
+  'Corrosive Projection',
+  'Dead Eye',
+  'EMP Aura',
+  'Empowered Blades',
+  'Enemy Radar',
+  'Energy Siphon',
+  'Growing Power',
+  'Infested Impedance',
+  'Loot Detector',
+  'Physique',
+  'Pistol Amp',
+  'Pistol Scavenger',
+  'Rejuvenation',
+  'Rifle Amp',
+  'Rifle Scavenger',
+  'Shield Disruption',
+  'Shotgun Amp',
+  'Shotgun Scavenger',
+  'Sniper Scavenger',
+  'Speed Holster',
+  'Sprint Boost',
+  'Stand United',
+  'Steel Charge',
+];
+
+// Nightmare mod names
+const nightmare = [
+  'Accelerated Blast',
+  'Animal Instinct',
+  'Armored Agility',
+  'Blaze',
+  'Chilling Reload',
+  'Constitution',
+  'Drifting Contact',
+  'Focus Energy',
+  'Fortitude',
+  'Hammer Shot',
+  'Ice Storm',
+  'Lethal Torrent',
+  'Rending Strike',
+  'Stunning Speed',
+  'Wildfire',
+  'Seeking Fury',
+  'Streamlined Form',
+  'Shred',
+  'Vigor',
+];
+
+module.exports = {
+  nightmare,
+  auras,
+  resources,
+};
